+++ src/data/mutationResults.ts
@@ -3,14 +3,14 @@ import {
 } from '../actions';
 
 // This is part of the public API, people write these functions in `updateQueries`.
-export type MutationQueryReducer = (previousResult: Object, options: {
-  mutationResult: Object,
+export type MutationQueryReducer<T> = (previousResult: Object, options: {
+  mutationResult: T | {},
   queryName: Object,
   queryVariables: Object,
 }) => Object;
 
-export type MutationQueryReducersMap = {
-  [queryName: string]: MutationQueryReducer;
+export type MutationQueryReducersMap<T> = {
+  [queryName: string]: MutationQueryReducer<T>;
 };
 
 export type OperationResultReducer = (previousResult: Object, action: ApolloAction, variables: Object) => Object;